Add files via upload
This is the Dhaka Subway Systems Automated Ticket Issuing System software development 
project management plan. The software development lifecycle is explained in detail in this 
software project management plan. This document will contain detailed information on the 
project's management plan. The people in the IT department are the intended audience for this 
project. It specifies the technical and managerial approaches to software product development. 
All engineering and management activities that need to be transferred to the Dhaka metro system.